backend/requirements.txt
asgiref==3.10.0
boto3==1.40.69
botocore==1.40.69
certifi==2025.10.5
charset-normalizer==3.4.4
dj-database-url==3.0.1
Django==5.0.14
django-cors-headers==4.9.0
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
jmespath==1.0.1
packaging==25.0
pillow==12.0.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.11.0
# AWS Cognito authentication dependencies
PyJWT==2.10.1
cryptography==44.0.0
# Stripe payment processing
stripe==12.0.0
# Google Maps geocoding
googlemaps==4.10.0
# TUI Admin Dashboard
textual>=0.47.0
rich>=13.7.0
# LLM for Instagram event extraction
anthropic>=0.18.0
